agent/requirements.txt
schedule
requests
beautifulsoup4
google-generativeai
python-dotenv
feedparser
